Speed Converter

+++ EPUBGenerator/EPUBGenerator/MainLogic/SoundEngine/PositionChangedEventArgs.cs
@@ -8,16 +8,16 @@ namespace EPUBGenerator.MainLogic.SoundEngine
 {
     public class PositionChangedEventArgs : EventArgs
     {
-        private long _change;
-        private long _position;
+        private double _change;
+        private double _position;
 
-        public long Change { get { return _change; } }
-        public long ByteChange { get { return _change * 2; } }
-        public long Position { get { return _position; } }
-        public long BytePosition { get { return _position * 2; } }
+        public double Change { get { return _change; } }
+        public double ByteChange { get { return _change * 2; } }
+        public double Position { get { return _position; } }
+        public double BytePosition { get { return _position * 2; } }
         public bool IsEnded { get; private set; }
 
-        public PositionChangedEventArgs(long change, long position, bool end)
+        public PositionChangedEventArgs(double change, double position, bool end)
         {
             _change = change;
             _position = position;
